Add update file parsing logic to a service and slim down the controller
