+++ spec/services/updates_parser_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe UpdatesController, type: :controller do
+RSpec.describe UpdatesParser do
   let(:date) { Date.new(2020, 04, 10) }
   let(:update_paths) do
     [
@@ -14,9 +14,9 @@ RSpec.describe UpdatesController, type: :controller do
     ]
   end
 
-  describe '#update_file_paths_to_hash' do
+  describe '#call' do
     it 'only adds valid update files to the hash' do
-      expect(subject.update_file_paths_to_hash(update_paths)[date])
+      expect(UpdatesParser.new(update_paths).call[date])
         .to eql([{ path: 'valid_update_title_2020_04_10', name: 'Valid update title' }])
     end
   end
